add modules (5.12)
