+++ js/data.js
@@ -1,0 +1,9 @@
+const BUILDINGS_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel']; //здания
+const CHECKIN_TIMES = ['12:00','13:00','14:00']; //время заезда
+const CHECKOUT_TIMES = ['12:00','13:00','14:00']; //время выезда
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS_PATH = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg'
+  ,'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg'
+  ,'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+
+export{BUILDINGS_TYPES, CHECKIN_TIMES, CHECKOUT_TIMES, FEATURES, PHOTOS_PATH};
